l06/requirements.txt
jupyter==1.0.0
matplotlib==3.2.0
networkx==2.4
numpy==1.18.1
pandas==1.0.3
pyro-ppl==1.3.0
scikit-learn==0.22.1
torch==1.4.0

